Reuploaded To Shorten Names - CW
Due to an error with GitHub desktop, my filenames and paths where too long for my snapshots.

John Busse was unable to push a commit on GitHub because of this issue. With this updated I have uploaded all my snapshots with shortened names.
The dates on them should still collateral with the date they were taken and committed.

This will also be explained in the txt uploaded.

Thank you for your patience - Cole Wheeler.